refactor: prize card 재사용 가능하도록 분리
(cherry picked from commit de65e007aa611ae9ebf0715a1c69cb6e36484967)
